Add ButtonColor updates for WSD colors
