+++ test/test_fixtures/rmui_project/pubspec.yaml
@@ -7,4 +7,4 @@ dependencies:
     hosted:
       name: react_material_ui
       url: https://pub.workiva.org
-    version: ^1.118.0
+    version: ^1.121.0
